feat.: Add ChatInfo to generated PDF
